Adding lightweight stream history.
Motivation:

The encoder and decoder have several cases that require information about streams that no longer exist in the connection. We currently have no way of obtaining information for these streams.

Modifications:

Added a history feature to Http2LifecycleManager that provides access to historical (i.e. removed) streams for some time period. Updated Http2ConnectionHandler to implement the history using a 2 map approach (new and old). When a periodic timer expires, the maps are swapped and we clear the oldest entries.

Updated the encoder and decoder to use the new history if the stream is not found in the connection.

Updated the local flow controller methods to support closed streams that have no flow state.  Also fixed a bug in the flow controller so that we now guarantee that all bytes will be consumed for closed streams.

Result:

Fixes #3643