Adding additional parameters to HTTP/2 examples
Motivation:

The current HTTP/2 example client/server do not allow customizing wither ALPN is used or forcing use of a particular SSL provider.  This would be helpful in testing various combinations of client/server support.

Modifications:

Modified the example client and server to provide parameters for whether to use ALPN, which SSL provider to use, as well as which protocols to advertise when using ALPN.

Result:

HTTP/2 examples are more configurable.